+++ src/bin/kxeslint-dump.js
@@ -12,4 +12,5 @@ if (!fs.existsSync(eslintFile)) {
 }
 const eslintConfig = (await import(eslintFile)).default;
 const JSON_INDENT = 2;
+
 console.log(JSON.stringify(eslintConfig, null, JSON_INDENT));