+++ src/bin/kxeslint-setup.js
@@ -4,10 +4,13 @@ import * as fs from "node:fs";
 import * as path from "node:path";
 import process from "node:process";
 
+import prompts from "prompts";
+
 import * as deps from "../dependencies.js";
 import {setDepCheck} from "../environ.js";
 
 const getEslintConfig = () => path.resolve(process.cwd(), "eslint.config.js");
+
 const depsCheck = async () => {
   console.group("eslint dependencies checking…");
   try {
@@ -43,22 +46,159 @@ const depsCheck = async () => {
     console.groupEnd();
   }
 };
+
 const eslintRcCandidates = [".js", ".cjs", ".json"];
+
 const getEslintrcPath = () => {
   for (const candidate of eslintRcCandidates) {
     const eslintrcPath = path.join(process.cwd(), `.eslintrc${candidate}`);
     if (fs.existsSync(eslintrcPath)) return eslintrcPath;
   }
 };
+
 const deleteLine = "DELETE THIS LINE WHEN CONFIGURED";
+
 const defaultConfig = `
 // ${deleteLine}
 import eslintConfig from "@keeex/eslint-config";
 
 export default await eslintConfig({});
 `.trim();
+
+const getOptionsFromTargetFilter = (target) => {
+  switch (target) {
+    case "js":
+      return {
+        haveMocha: "maybe",
+        haveReact: false,
+      };
+    case "node":
+      return {
+        haveMocha: "maybe",
+        haveReact: "maybe",
+      };
+    case "browser":
+      return {
+        haveMocha: "maybe",
+        haveReact: "maybe",
+      };
+    case "react-native":
+      return {
+        haveMocha: false,
+        haveReact: true,
+      };
+  }
+  throw new Error("Unexpected state");
+};
+
+const getBooleanPrompt = async (message) => {
+  const replyValue = await prompts({
+    initial: true,
+    message,
+    name: "value",
+    type: "toggle",
+  });
+  return replyValue.value;
+};
+
+const getOptionsFromTarget = async (target) => {
+  let {haveMocha, haveReact} = await getOptionsFromTargetFilter(target);
+  if (haveReact === "maybe") haveReact = await getBooleanPrompt("Use React?");
+  if (haveMocha === "maybe") haveMocha = await getBooleanPrompt("Use Mocha?");
+  return {haveReact, haveMocha};
+};
+
+const getAvailableEnv = (target, haveMocha, haveReact) => {
+  switch (target) {
+    case "js":
+      if (haveMocha) return ["library"];
+      return [];
+    case "node":
+      if (haveReact) return ["webservice", "node", "none"];
+      return ["node"];
+    case "browser":
+      return ["webapp", "weblibrary"];
+    case "react-native":
+      return ["mobile"];
+  }
+  throw new Error("Unexpected state");
+};
+
+const getEnvTitle = (envName) => {
+  switch (envName) {
+    case "webservice":
+      return "Webservice (server + webapp)";
+    case "webapp":
+      return "Webapp only";
+    case "weblibrary":
+      return "Library for browsers only";
+    case "mobile":
+      return "React-native mobile app";
+    case "library":
+      return "Pure JavaScript library";
+    case "node":
+      return "NodeJS application";
+    case "none":
+      return "Do not pick an environment";
+  }
+  throw new Error("Unexpected state");
+};
+
+const getEnvFromTarget = async (target, haveMocha, haveReact) => {
+  const availableEnv = getAvailableEnv(target, haveMocha, haveReact);
+  if (availableEnv.length === 0) return null;
+  if (availableEnv.length === 1) return availableEnv[0];
+  const choice = await prompts({
+    choices: availableEnv.map((e) => ({title: getEnvTitle(e), value: e})),
+    message: "Target environment?",
+    name: "value",
+    type: "select",
+  });
+  if (!choice.value) return null;
+  return choice.value;
+};
+
+const genConfig = (targetEnv, mocha, react) => {
+  const configRows = [];
+  if (targetEnv && targetEnv !== "none") {
+    configRows.push(`"environments": ${JSON.stringify(targetEnv)}`);
+  }
+  configRows.push(`"mocha": ${JSON.stringify(mocha)}`);
+  configRows.push(`"react": ${JSON.stringify(react)}`);
+  configRows.push('"typescript": true');
+  return `
+import eslintConfig from "@keeex/eslint-config";
+
+export default await eslintConfig({
+${configRows.map((c) => `  ${c},`).join("\n")}
+});
+`.trim();
+};
+
+/** Generate a new base config based on a few questions. */
+const getNewConfig = async () => {
+  const generalTarget = await prompts({
+    choices: [
+      {title: "Generic JavaScript", value: "js"},
+      {title: "NodeJS", value: "node"},
+      {title: "Browser", value: "browser"},
+      {title: "React Native", value: "react-native"},
+    ],
+    message: "General target environment",
+    name: "value",
+    type: "select",
+  });
+  const target = generalTarget.value;
+  if (!target) return null;
+  const {haveMocha, haveReact} = await getOptionsFromTarget(target);
+  const targetEnv = await getEnvFromTarget(target, haveMocha, haveReact);
+  return genConfig(targetEnv, haveMocha, haveReact);
+};
+
+const enforceNewline = (str) => (str.endsWith("\n") ? str : `${str}\n`);
+
 /** Return true if the flat config system is ok */
-const checkEslintConfig = () => {
+const checkEslintConfig = async () => {
   console.group("Checking eslint config file…");
   try {
     const eslintrcPath = getEslintrcPath();
@@ -77,11 +217,15 @@ const checkEslintConfig = () => {
       return false;
     }
     if (!fs.existsSync(eslintConfigPath)) {
-      console.log(
-        `- No ${eslintConfigPath} found. Generating empty config. Please configure it appropriately.`,
-      );
-      fs.writeFileSync(eslintConfigPath, `${defaultConfig}\n`);
-      return false;
+      console.log(`- No ${eslintConfigPath} found. Generating new config.`);
+      const configContent = await getNewConfig();
+      if (configContent === null) {
+        console.log("  Default config file generated. You'll have to edit it manually.");
+        fs.writeFileSync(eslintConfigPath, enforceNewline(defaultConfig));
+        return false;
+      }
+      fs.writeFileSync(eslintConfigPath, enforceNewline(configContent));
+      return true;
     }
     const configContent = fs.readFileSync(eslintConfigPath, "utf8");
     if (configContent.indexOf(deleteLine) !== -1) {
@@ -99,7 +243,9 @@ const checkEslintConfig = () => {
     console.groupEnd();
   }
 };
+
 const PRETTIER_CONFIG = "prettier.config.js";
+
 const prettierSetup = () => {
   console.group("Checking prettier config");
   try {
@@ -120,6 +266,7 @@ export default config;
     console.groupEnd();
   }
 };
+
 const checkTsConfig = () => {
   const needTsConfig = deps.needDependencies("typescript-eslint");
   if (!needTsConfig) return;
@@ -132,10 +279,22 @@ const checkTsConfig = () => {
     console.groupEnd();
   }
 };
+
+const checkProjectType = () => {
+  const packageContent = JSON.parse(fs.readFileSync("package.json", "utf8"));
+  if (packageContent.type !== "module") {
+    console.error("package.json type is not module.");
+    return false;
+  }
+  return true;
+};
+
 const main = async () => {
   console.group("KeeeX eslint and prettier configuration setup");
   try {
-    const configOk = checkEslintConfig();
+    const typeOk = checkProjectType();
+    if (!typeOk) return;
+    const configOk = await checkEslintConfig();
     if (!configOk) return;
     await prettierSetup();
     await depsCheck();