+++ src/environ.js
@@ -2,5 +2,6 @@
 export const setDepCheck = () => {
   process.env.KEEEX_DEPCHECK = "true";
 };
+
 /** Check the depcheck flag */
 export const isDepCheck = () => process.env.KEEEX_DEPCHECK === "true";